refactor: extract event duration utility
